feature: Add support for multithreaded requests and logback

+++ src/main/java/com/github/sinakarimi81/jdown/dataObjects/Range.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Range {
 
-    private long from;
-    private long to;
+    private int from;
+    private int to;
 
 }
